:bug: Fix confusing StampRally and Diagnose story
