+++ components/toppage/Feature.stories.tsx
@@ -20,18 +20,14 @@ const featureContentStyle = css`
   width: 100%;
 `;
 
-const stampRallyContent = (
-  <img
-    src="/toppage/stamp-rally.png"
-    alt="スタンプラリー"
-    css={featureContentStyle}
-  />
+const diagnoseContent = (
+  <img src="/toppage/diagnose.png" alt="相性診断" css={featureContentStyle} />
 );
 
-export const StampRally = Template.bind({});
-StampRally.args = {
-  title: "スタンプラリー",
-  description: "QRコードを読み込んで、景品を貰いに行こう！",
-  featureContentNode: stampRallyContent,
-  inverse: true,
+export const Diagnose = Template.bind({});
+Diagnose.args = {
+  title: "相性診断",
+  description:
+    "BINGOスタンプラリーのために相性診断をして自分に合った部・サークルの説明を聞きに行こう！",
+  featureContentNode: diagnoseContent,
 };
